Move error macros to detail + move traits to utility
