+++ include/cuco/detail/bitwise_compare.cuh
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2022, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include <cuco/traits.hpp>
+#include <cuco/utility/traits.hpp>
 
 #include <cstdint>
 #include <type_traits>
