+++ include/cuco/dynamic_map.cuh
@@ -20,7 +20,6 @@
 #include <cuco/hash_functions.cuh>
 #include <cuco/sentinel.cuh>
 #include <cuco/static_map.cuh>
-#include <cuco/utility/error.hpp>
 
 #include <thrust/device_vector.h>
 #include <thrust/functional.h>
