+++ include/cuco/static_multimap.cuh
@@ -21,9 +21,8 @@
 #include <cuco/hash_functions.cuh>
 #include <cuco/probe_sequences.cuh>
 #include <cuco/sentinel.cuh>
-#include <cuco/traits.hpp>
 #include <cuco/utility/allocator.hpp>
-#include <cuco/utility/error.hpp>
+#include <cuco/utility/traits.hpp>
 
 #include <thrust/functional.h>
